--- CONTRIBUTING.md
@@ -1,57 +1,0 @@
-# Contributing to lichess-bot
-
-We welcome your contributions! There are multiple ways to contribute.
-
-## Table of Contents
-
-1. [Code of Conduct](#code-of-conduct)
-2. [How to Contribute](#how-to-contribute)
-3. [Reporting Bugs](#reporting-bugs)
-4. [Requesting Features](#requesting-features)
-5. [Submitting Pull Requests](#submitting-pull-requests)
-6. [Testing](#testing)
-7. [Documentation](#documentation)
-
-## Code of Conduct
-
-Please review our [Code of Conduct](/CODE_OF_CONDUCT.md) before participating in our community. We want all contributors to feel welcome and to foster an open and inclusive environment.
-
-## How to Contribute
-
-We welcome contributions in the form of bug reports, feature requests, code changes, and documentation improvements. Here's how you can contribute:
-
-- Fork the repository to your GitHub account.
-- Create a new branch for your feature or bug fix.
-- Make your changes and commit them with a clear and concise commit message.
-- Push your changes to your branch.
-- Submit a pull request to the main repository.
-
-Please follow our [Pull Request Template](.github/pull_request_template.md) when submitting a pull request.
-
-## Reporting Bugs
-
-If you find a bug, please open an issue with a detailed description of the problem. Include information about your environment and steps to reproduce the issue.
-When filing a bug remember that the better written the bug is, the more likely it is to be fixed.
-Please follow our [Bug Report Template](.github/ISSUE_TEMPLATE/bug_report.md) when submitting a pull request.
-
-## Requesting Features
-
-We encourage you to open an issue to propose new features or improvements. Please provide as much detail as possible about your suggestion.
-Please follow our [Feature Request Template](.github/ISSUE_TEMPLATE/feature_request.md) when submitting a pull request.
-
-## Submitting Pull Requests
-
-When submitting a pull request, please ensure the following:
-
-- You have added or updated relevant documentation.
-- Tests (if applicable) have been added or updated.
-- Your branch is up-to-date with the main repository.
-- The pull request title and description are clear and concise.
-
-## Testing
-
-Ensure that your changes pass all existing tests and consider adding new tests if applicable.
-
-## Documentation
-
-Improvements to the documentation are always welcome. If you find areas that need clarification or additional information, please submit a pull request.
