+++ lib/strategies.py
